Replacing byte* in NLSF codebook with short*

+++ CSharp/Concentus/Silk/DecodeIndices.cs
@@ -52,7 +52,7 @@ namespace Concentus.Silk
             int i, k, Ix;
             int decode_absolute_lagIndex, delta_lagIndex;
             short[] ec_ix = new short[psDec.LPC_order];
-            byte[] pred_Q8 = new byte[psDec.LPC_order];
+            short[] pred_Q8 = new short[psDec.LPC_order];
 
             /*******************************************/
             /* Decode signal type and quantizer offset */
