+++ CSharp/Concentus/Silk/EncodeIndices.cs
@@ -58,7 +58,7 @@ namespace Concentus.Silk
             int i, k, typeOffset;
             int encode_absolute_lagIndex, delta_lagIndex;
             short[] ec_ix = new short[SilkConstants.MAX_LPC_ORDER];
-            byte[] pred_Q8 = new byte[SilkConstants.MAX_LPC_ORDER];
+            short[] pred_Q8 = new short[SilkConstants.MAX_LPC_ORDER];
             SideInfoIndices psIndices;
 
             if (encode_LBRR != 0)
